Rename endpoint module to service and remove service stop function
