+++ component/src/main/wso2/kafka/consumer.bal
@@ -138,7 +138,7 @@ public type SimpleConsumer client object {
     }
 
     public function __stop() returns error? {
-        return self.stop();
+        return self->close();
     }
 
     public function __attach(service s, map<any> annotationData) returns error? {
@@ -150,8 +150,6 @@ public type SimpleConsumer client object {
     extern function registerListener(service serviceType, map<any> annotationData) returns error?;
 
     extern function start() returns error?;
-
-    extern function stop() returns error?;
 
     # Assigns consumer to a set of topic partitions.
     #
